+++ src/Kephas.Extensions.DependencyInjection/release-notes.md
@@ -7,3 +7,4 @@
 
 * NEW: Added ```serviceCollection.AddInjectionExtensions()``` extension method to provide the infrastructure for Kephas integration with the ```Microsoft.Extensions.DependencyInjection```.
 * NEW: Added ```README.md``` and ```release-notes.md```.
+* Breaking change: `IAmbientServices.WithServiceCollection(services)` refactored to `IServiceCollection.UseAmbientServices(ambientServices)`.